ai-agent/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.7
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.1
pinecone-client==2.2.4
weaviate-client==3.25.3
chromadb==0.4.18
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
aiohttp==3.9.1
asyncio==3.4.3
redis==5.0.1
celery==5.3.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pillow==10.1.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
tiktoken==0.5.1
tqdm==4.66.1
loguru==0.7.2
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1 